refactor(rpc): extract read plan executor
Move TX_EXECUTE_READ_PLAN execution and plan-key binding helpers out of lineairdb_rpc.cc into read_plan_executor.cc.
Register read_plan_executor.cc in the server build.
